+++ .gitignore
@@ -7,3 +7,5 @@ build/*
 .externalNativeBuild
 .cxx
 local.properties
+.lock
+.bin
